# Conflicts:
#	Backend/src/main/java/com/example/backend/project/controller/ProjectController.java
#	Backend/src/main/java/com/example/backend/project/service/ProjectService.java
#	Backend/src/main/java/com/example/backend/user/service/UserService.java
#	Frontend/src/components/pages/Project.jsx

+++ Backend/src/main/java/com/example/backend/project/service/ProjectService.java
@@ -2,6 +2,8 @@ package com.example.backend.project.service;
 
 import com.example.backend.collaborations.model.Collaboration;
 import com.example.backend.collaborations.repo.CollaborationsRepository;
+import com.example.backend.project.controller.dpo.FRTeamMemberDPO;
+import com.example.backend.project.controller.dpo.ProjectAndFRTeamMembersDPO;
 import com.example.backend.companies.model.Company;
 import com.example.backend.project.controller.dto.ProjectDTO;
 import com.example.backend.project.model.Project;
@@ -9,15 +11,17 @@ import com.example.backend.project.repo.ProjectRepository;
 import com.example.backend.user.model.AUTHORITY;
 import com.example.backend.user.model.AppUser;
 import com.example.backend.user.repo.UserRepository;
+import com.example.backend.util.exceptions.EntityNotFoundException;
 import lombok.AllArgsConstructor;
-import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.naming.AuthenticationException;
 import javax.persistence.EntityNotFoundException;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Optional;
+import java.util.Objects;
+import java.util.Set;
 
 @Service @Transactional
 @AllArgsConstructor
@@ -26,39 +30,56 @@ public class ProjectService {
     private UserRepository userRepository;
     private CollaborationsRepository collaborationsRepository;
 
-    public List<Project> findAll() {
+    public List<Project> findAll(AppUser user) throws AuthenticationException {
+        if (user == null) throw new AuthenticationException();
+
         return projectRepository.findAll();
     }
 
-    public Project findById(Long id) {
-        return projectRepository.findById(id).get();
+    public ProjectAndFRTeamMembersDPO findById(Long id, AppUser user) throws AuthenticationException, EntityNotFoundException {
+        if (user == null) throw new AuthenticationException("You do not have permission to access CDB.");
+
+        Project project;
+        if (projectRepository.findById(id).isPresent()) project = projectRepository.findById(id).get();
+        else throw new EntityNotFoundException("You do not have permission to execute this command.");
+        Set<FRTeamMemberDPO> frTeamMemberDPOS = new HashSet<>();
+        for (AppUser appUser: project.getFrteammembers()){
+            FRTeamMemberDPO x = new FRTeamMemberDPO(
+                appUser.getId(),
+                appUser.getFirstName(),
+                appUser.getLastName()
+            );
+            frTeamMemberDPOS.add(x);
+        }
+        return new ProjectAndFRTeamMembersDPO(project, frTeamMemberDPOS);
     }
 
-    public Project addProject(ProjectDTO projectDTO) {
-        AppUser user = userRepository.findById(projectDTO.getIdFRResp()).get();
-        if (user.getAuthority() == AUTHORITY.OBSERVER || user.getAuthority() == AUTHORITY.FR_TEAM_MEMBER){
-            user.setAuthority(AUTHORITY.FR_RESPONSIBLE);
-            userRepository.save(user);
+    public Project addProject(ProjectDTO projectDTO, AppUser user) throws EntityNotFoundException, AuthenticationException {
+        if (user == null) throw new AuthenticationException("You do not have permission to access CDB.");
+        if (!List.of(AUTHORITY.ADMINISTRATOR, AUTHORITY.MODERATOR).contains(user.getAuthority())) throw new AuthenticationException();
+
+        AppUser frResp;
+        if (userRepository.findById(projectDTO.getIdFRResp()).isPresent()) frResp = userRepository.findById(projectDTO.getIdFRResp()).get();
+        else throw new EntityNotFoundException();
+
+        if (frResp.getAuthority() == AUTHORITY.OBSERVER || frResp.getAuthority() == AUTHORITY.FR_TEAM_MEMBER){
+            frResp.setAuthority(AUTHORITY.FR_RESPONSIBLE);
+            userRepository.save(frResp);
         }
-        Project project = new Project(
-                projectDTO.getIdCreator(),
-                projectDTO.getName(),
-                projectDTO.getCategory(),
-                projectDTO.getType(),
-                projectDTO.getStartDate(),
-                projectDTO.getEndDate(),
-                user,
-                new HashSet<>(),
-                projectDTO.getFRgoal(),
-                projectDTO.getFirstPingDate(),
-                projectDTO.getSecondPingDate()
-        );
+        Project project = projectDTO.toProject();
+        project.setFRResp(frResp);
         project.addFrTeamMember(user);
         return projectRepository.save(project);
     }
 
-    public Project updateProject(ProjectDTO projectDTO, Long id) {
-        Project project = projectRepository.findById(id).get();
+    public Project updateProject(ProjectDTO projectDTO, Long id, AppUser user) throws AuthenticationException, EntityNotFoundException {
+        if (user == null) throw new AuthenticationException("You do not have permission to access CDB.");
+        if (!List.of(AUTHORITY.ADMINISTRATOR, AUTHORITY.MODERATOR).contains(user.getAuthority())) throw new AuthenticationException("You do not have permission to execute this command.");
+
+
+        Project project;
+        if (projectRepository.findById(id).isPresent()) project = projectRepository.findById(id).get();
+        else throw new EntityNotFoundException("User under id " + id + " not found.");
 
         project.setIdCreator(projectDTO.getIdCreator());
         project.setName(projectDTO.getName());
@@ -66,7 +87,8 @@ public class ProjectService {
         project.setType(projectDTO.getType());
         project.setStartDate(projectDTO.getStartDate());
         project.setEndDate(projectDTO.getEndDate());
-        project.setFRResp(userRepository.findById(projectDTO.getIdFRResp()).get());
+        if (userRepository.findById(projectDTO.getIdFRResp()).isPresent()) project.setFRResp(userRepository.findById(projectDTO.getIdFRResp()).get());
+        else throw new EntityNotFoundException();
         project.setFRGoal(projectDTO.getFRgoal());
         project.setFirstPingDate(projectDTO.getFirstPingDate());
         project.setSecondPingDate(projectDTO.getSecondPingDate());
@@ -74,20 +96,45 @@ public class ProjectService {
         return projectRepository.save(project);
     }
 
-    public List<Collaboration> getCollaborations(Long id) {
-        Project project = projectRepository.findById(id).get();
+    public List<Collaboration> getCollaborations(Long id, AppUser user) throws AuthenticationException, EntityNotFoundException {
+        if (user == null) throw new AuthenticationException("You do not have permission to access CDB.");
+
+        AUTHORITY authority = user.getAuthority();
+        Project project;
+        if (projectRepository.findById(id).isPresent()) project = projectRepository.findById(id).get();
+        else throw new EntityNotFoundException("Project under id " + id + " not found.");
+
+        if (authority == AUTHORITY.OBSERVER ||
+                (authority == AUTHORITY.FR_TEAM_MEMBER && !project.getFrteammembers().contains(user)) ||
+                (authority == AUTHORITY.FR_RESPONSIBLE && !Objects.equals(project.getFRResp().getId(), user.getId())))
+            throw new AuthenticationException("You do not have permission to execute this command.");
+
         return collaborationsRepository.findAllByCollaborationId_Project(project);
     }
 
-    public void addFrTeamMember(Long projectId, Long teamMemberId){
-        Project project = projectRepository.findById(projectId).get();
-        AppUser user = userRepository.findById(teamMemberId).get();
-        if (user.getAuthority() == AUTHORITY.OBSERVER) {
+    public Project addFrTeamMember(Long projectId, Long teamMemberId, AppUser user) throws AuthenticationException, EntityNotFoundException {
+        if (user == null) throw new AuthenticationException("You do not have permission to access CDB.");
+        AUTHORITY authority = user.getAuthority();
+        if (!List.of(AUTHORITY.ADMINISTRATOR, AUTHORITY.MODERATOR, AUTHORITY.FR_RESPONSIBLE).contains(authority))
+            throw new AuthenticationException("You do not have permission to execute this command.");
+        Project project;
+        if (projectRepository.findById(projectId).isPresent()) project = projectRepository.findById(projectId).get();
+        else throw new EntityNotFoundException("Project under id " + projectId + " not found.");
+        AppUser newFRTeamMember;
+        if (userRepository.findById(teamMemberId).isPresent()) newFRTeamMember = userRepository.findById(teamMemberId).get();
+        else throw new EntityNotFoundException("User under id " + teamMemberId + " not found.");
+
+        if (authority == AUTHORITY.FR_RESPONSIBLE && !Objects.equals(project.getFRResp().getId(), user.getId())) {
+            throw new AuthenticationException("You do not have permission to execute this command.");
+        }
+
+        if (newFRTeamMember.getAuthority() == AUTHORITY.OBSERVER) {
             user.setAuthority(AUTHORITY.FR_TEAM_MEMBER);
             userRepository.save(user);
         }
         project.addFrTeamMember(user);
-        projectRepository.save(project);
+
+        return projectRepository.save(project);
     }
 
     public Boolean softLockProject(Long id) throws EntityNotFoundException
@@ -99,18 +146,30 @@ public class ProjectService {
         return newSoftLock;
     }
 
-    public void deleteFrTeamMember(Long projectId, Long teamMemberId){
-        Project project = projectRepository.findById(projectId).get();
-        AppUser user = userRepository.findById(teamMemberId).get();
-        project.removeFrTeamMember(user);
-        if (user.getProjects().isEmpty() && user.getAuthority() == AUTHORITY.FR_TEAM_MEMBER){
-            user.setAuthority(AUTHORITY.OBSERVER);
-            userRepository.save(user);
+    public Project deleteFrTeamMember(Long projectId, Long teamMemberId, AppUser user) throws AuthenticationException, EntityNotFoundException {
+        if (user == null) throw new AuthenticationException("You do not have permission to access CDB.");
+
+        Project project;
+        if (projectRepository.findById(projectId).isPresent()) project = projectRepository.findById(projectId).get();
+        else throw new EntityNotFoundException("Project under id " + projectId + " not found.");
+        AppUser FRTeamMember;
+        if (userRepository.findById(teamMemberId).isPresent()) FRTeamMember = userRepository.findById(teamMemberId).get();
+        else throw new EntityNotFoundException("User under id " + teamMemberId + " not found.");
+
+        project.removeFrTeamMember(FRTeamMember);
+        if (FRTeamMember.getProjects().isEmpty() && FRTeamMember.getAuthority() == AUTHORITY.FR_TEAM_MEMBER){
+            FRTeamMember.setAuthority(AUTHORITY.OBSERVER);
+            userRepository.save(FRTeamMember);
         }
-        projectRepository.save(project);
+        return projectRepository.save(project);
     }
 
-    public void deleteProject(Long id) {
+    public void deleteProject(Long id, AppUser user) throws AuthenticationException, EntityNotFoundException {
+        if (user == null) throw new AuthenticationException("You do not have permission to access CDB.");
+        if (!List.of(AUTHORITY.ADMINISTRATOR, AUTHORITY.MODERATOR).contains(user.getAuthority())) throw new AuthenticationException("You do not have permission to execute this command.");
+
+        if (!projectRepository.existsById(id)) throw new EntityNotFoundException("Project under id " + id + " not found.");
+
         projectRepository.deleteById(id);
     }
